Refactor: Implement HTML rendering in LanguageContext
This commit implements the plan to allow HTML tags to be rendered correctly within translated strings. It adds a `tHTML` function to `LanguageContext.tsx` that returns JSX elements and updates the `Transformation` component to use this function for rendering the title and subtitle, enabling the use of `<br>` and `<em>` tags.